Mitad de DTOAssembler hechos

+++ Demo/src/main/java/co/edu/co/extraclase/business/domain/PriorityDomain.java
@@ -63,7 +63,7 @@ public final class PriorityDomain extends Domain {
 		this.description = TextHelper.getDefaultWithTrim(description);
 	}
 
-	public float getResponseTime() {
+	public Integer getResponseTime() {
 		return responseTime;
 	}
